+++ lib/src/channel/axeptio_sdk_platform_interface.dart
@@ -64,6 +64,26 @@ abstract interface class AxeptioSdkPlatform extends PlatformInterface {
     throw UnimplementedError('getConsentSavedData() has not been implemented');
   }
 
+  Future<Map<String, dynamic>?> getConsentDebugInfo({String? preferenceKey}) {
+    throw UnimplementedError('getConsentDebugInfo() has not been implemented');
+  }
+
+  Future<Map<int, bool>> getVendorConsents() {
+    throw UnimplementedError('getVendorConsents() has not been implemented');
+  }
+
+  Future<List<int>> getConsentedVendors() {
+    throw UnimplementedError('getConsentedVendors() has not been implemented');
+  }
+
+  Future<List<int>> getRefusedVendors() {
+    throw UnimplementedError('getRefusedVendors() has not been implemented');
+  }
+
+  Future<bool> isVendorConsented(int vendorId) {
+    throw UnimplementedError('isVendorConsented() has not been implemented');
+  }
+
   addEventListener(AxeptioEventListener listener) {
     throw UnimplementedError('addEventListener() has not been implemented');
   }