+++ test/axeptio_sdk_test.dart
@@ -69,6 +69,36 @@ class MockAxeptioSdkPlatform
     // TODO: implement getConsentSavedData
     throw UnimplementedError();
   }
+
+  @override
+  Future<Map<String, dynamic>?> getConsentDebugInfo({String? preferenceKey}) {
+    // TODO: implement getConsentDebugInfo
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<Map<int, bool>> getVendorConsents() {
+    // TODO: implement getVendorConsents
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<List<int>> getConsentedVendors() {
+    // TODO: implement getConsentedVendors
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<List<int>> getRefusedVendors() {
+    // TODO: implement getRefusedVendors
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<bool> isVendorConsented(int vendorId) {
+    // TODO: implement isVendorConsented
+    throw UnimplementedError();
+  }
 }
 
 void main() {